requirements.txt
requests>=2.31.0
openai>=1.92.0
discord.py>=2.5.2 